requirements.txt
django==5.0.*
django-allauth~=0.60
psycopg[binary]~=3.1
python-environ==0.4.54
celery~=5.3
redis~=5.0
requests~=2.32
weasyprint~=60.2
cryptography~=42.0
python-dotenv~=1.0